feat: implement redeem points and withdraw request functionalities with UI updates

+++ app/src/main/java/com/ovasta/sellers/presentation/profile/wallet/data/WithdrawRequests.kt
@@ -2,10 +2,9 @@ package com.ovasta.sellers.presentation.profile.wallet.data
 
 import com.google.gson.annotations.SerializedName
 
-
 data class WithdrawRequests(
     @SerializedName("id") var id: Int,
-    @SerializedName("amount") var amount: Int? = null,
+    @SerializedName("amount") var amount: Double? = null,
     @SerializedName("status") var status: Int,
     @SerializedName("rejection_reason") var rejectionReason: String? = null,
     @SerializedName("created_at") var createdAt: String? = null,
